pyproject.toml
[tool.poetry]
name = "ruble-tracker"
version = "0.1.0"
description = ""
authors = ["Алексей Милов <killerarz@yandex.ru>"]
readme = "README.md"
#packages = [{include = "ruble_tracker"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.3.0"
celery = "^5.3.4"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
